backend/requirements.txt
# Core Dependencies
streamlit>=1.42.0
python-dotenv==1.0.1

# LangChain & LLM
langchain==0.1.16
langchain-groq==0.1.3
langchain-community==0.0.34
langchain-core==0.1.45
langchain-experimental==0.0.52

# Data Processing
pandas>=2.2.0
numpy>=1.24.0,<2.0
openpyxl>=3.1.2
xlrd>=2.0.1

# Statistical Analysis
scipy>=1.11.0
scikit-learn>=1.4.0
statsmodels>=0.14.0

# ML & Explainability
xgboost>=2.0.0
shap>=0.45.0

# Agentic System (NEW)
langgraph>=0.0.20
prophet>=1.1.5

# Visualization
plotly==5.20.0
matplotlib==3.8.3
seaborn==0.13.2

# RAG & Embeddings (NEW)
chromadb>=0.5.0
sentence-transformers>=2.5.1
tiktoken>=0.6.0
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-api
opentelemetry-sdk
mmh3
pypika
posthog<6.0.0
importlib-resources
onnxruntime
overrides

# Document Processing (NEW)
pypdf==4.0.1
python-docx==1.1.0
markdown==3.5.2

# Tools & Utilities (NEW)
duckduckgo-search==5.0.0
wikipedia==1.4.0

# Export Tools (NEW)
reportlab==4.1.0
xlsxwriter==3.2.0

# Caching & Performance (NEW)
diskcache==5.6.3

# Authentication & Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
bcrypt>=4.0.1
pyjwt>=2.8.0

# Testing (NEW)
pytest==8.0.0
pytest-asyncio==0.23.5

# Cloud Storage (NEW)
google-auth==2.28.1
google-api-python-client==2.126.0
msal==1.27.0

# Scheduling (NEW)
apscheduler==3.10.4

# Chart export (NEW)
kaleido==0.2.1
